+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/info_panel/util/constants.ts
@@ -10,7 +10,3 @@ export const ASSET_TYPE = {
 	FOLDER: 'folder',
 	KNOWLEDGE_BASES: 'knowledge-bases',
 };
-
-export const EVENTS = {
-	ASSET_DATA: 'asset-data',
-};
