+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/props_transformer/AllFDSPropsTransformer.tsx
@@ -4,30 +4,29 @@
  */
 
 import {IInternalRenderer, IView} from '@liferay/frontend-data-set-web';
+import {openModal} from 'frontend-js-components-web';
+import React from 'react';
 
+import formatActionURL from '../../common/utils/formatActionURL';
+import {ISearchAssetObjectEntry} from '../../structure_builder/types/AssetType';
 import AssetTypeInfoPanel from '../info_panel/AssetTypeInfoPanelContent';
-import {EVENTS} from '../info_panel/util/constants';
+import FilePreviewerModalContent from '../modal/FilePreviewerModalContent';
 import createAssetAction from './actions/createAssetAction';
-import createFolderAction from './actions/createFolderAction';
 import multipleFilesUploadAction from './actions/multipleFilesUploadAction';
 import shareAction from './actions/shareAction';
 import AuthorRenderer from './cell_renderers/AuthorRenderer';
 import NameRenderer from './cell_renderers/NameRenderer';
-import SimpleActionLinkRenderer from './cell_renderers/SimpleActionLinkRenderer';
+import SpaceRenderer from './cell_renderers/SpaceRenderer';
 import TypeRenderer from './cell_renderers/TypeRenderer';
 import addOnClickToCreationMenuItems from './utils/addOnClickToCreationMenuItems';
 import transformViewsItemsProps from './utils/transformViewsItemProps';
 
 const ACTIONS = {
 	createAsset: createAssetAction,
-	createFolder: createFolderAction,
 	uploadMultipleFiles: multipleFilesUploadAction,
 };
 
-const OBJECT_ENTRY_FOLDER_CLASSNAME =
-	'com.liferay.object.model.ObjectEntryFolder';
-
-export default function FolderFDSPropsTransformer({
+export default function AllFDSPropsTransformer({
 	additionalProps,
 	creationMenu,
 	itemsActions = [],
@@ -66,8 +65,8 @@ export default function FolderFDSPropsTransformer({
 					type: 'internal',
 				} as IInternalRenderer,
 				{
-					component: SimpleActionLinkRenderer,
-					name: 'simpleActionLinkTableCellRenderer',
+					component: SpaceRenderer,
+					name: 'spaceTableCellRenderer',
 					type: 'internal',
 				} as IInternalRenderer,
 				{
@@ -77,7 +76,9 @@ export default function FolderFDSPropsTransformer({
 				} as IInternalRenderer,
 			],
 		},
-		infoPanelComponent: () => AssetTypeInfoPanel({additionalProps}),
+		infoPanelComponent: (items: {items: ISearchAssetObjectEntry[]}) => (
+			<AssetTypeInfoPanel {...(otherProps as any)} {...items} />
+		),
 		itemsActions: itemsActions.map((action) => {
 			if (action?.data?.id === 'download') {
 				return {
@@ -86,14 +87,16 @@ export default function FolderFDSPropsTransformer({
 						Boolean(item?.embedded?.file?.link?.href),
 				};
 			}
-			else if (action?.data?.id === 'actionLink') {
+			else if (action?.data?.id === 'view-content') {
 				return {
 					...action,
-					isVisible: (item: any) =>
-						Boolean(
-							item?.entryClassName !==
-								OBJECT_ENTRY_FOLDER_CLASSNAME
-						),
+					isVisible: (item: any) => Boolean(!item?.embedded?.file),
+				};
+			}
+			else if (action?.data?.id === 'view-file') {
+				return {
+					...action,
+					isVisible: (item: any) => Boolean(item?.embedded?.file),
 				};
 			}
 
@@ -101,13 +104,34 @@ export default function FolderFDSPropsTransformer({
 		}),
 		onActionDropdownItemClick: ({
 			action,
+			event,
 			itemData,
 		}: {
 			action: any;
+			event: Event;
 			itemData: any;
 		}) => {
-			if (action?.data?.id === 'show-details') {
-				Liferay.fire(EVENTS.ASSET_DATA, {items: [{...itemData}]});
+			if (action?.data?.id === 'view-content') {
+				event?.preventDefault();
+
+				openModal({
+					size: 'full-screen',
+					title: itemData.embedded.title,
+					url: formatActionURL(itemData, action.href),
+				});
+			}
+			else if (action?.data?.id === 'view-file') {
+				openModal({
+					containerProps: {
+						className: '',
+					},
+					contentComponent: () =>
+						FilePreviewerModalContent({
+							file: itemData.embedded.file,
+							headerName: itemData.embedded.title,
+						}),
+					size: 'full-screen',
+				});
 			}
 			else if (action?.data?.id === 'share') {
 				const {autocompleteURL, collaboratorURLs} = additionalProps;
@@ -120,9 +144,6 @@ export default function FolderFDSPropsTransformer({
 					title: itemData.embedded?.title,
 				});
 			}
-		},
-		onSelectedItemsChange: (selectedItems: any[]) => {
-			Liferay.fire(EVENTS.ASSET_DATA, {items: selectedItems});
 		},
 		views: transformViewsItemsProps(views),
 	};
