+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/props_transformer/ContentsFDSPropsTransformer.tsx
@@ -5,16 +5,17 @@
 
 import {IInternalRenderer, IView} from '@liferay/frontend-data-set-web';
 import {openModal} from 'frontend-js-components-web';
+import React from 'react';
 
 import formatActionURL from '../../common/utils/formatActionURL';
+import {ISearchAssetObjectEntry} from '../../structure_builder/types/AssetType';
 import AssetTypeInfoPanel from '../info_panel/AssetTypeInfoPanelContent';
-import {EVENTS} from '../info_panel/util/constants';
-import FilePreviewerModalContent from '../modal/FilePreviewerModalContent';
 import createAssetAction from './actions/createAssetAction';
-import multipleFilesUploadAction from './actions/multipleFilesUploadAction';
+import createFolderAction from './actions/createFolderAction';
 import shareAction from './actions/shareAction';
 import AuthorRenderer from './cell_renderers/AuthorRenderer';
 import NameRenderer from './cell_renderers/NameRenderer';
+import SimpleActionLinkRenderer from './cell_renderers/SimpleActionLinkRenderer';
 import SpaceRenderer from './cell_renderers/SpaceRenderer';
 import TypeRenderer from './cell_renderers/TypeRenderer';
 import addOnClickToCreationMenuItems from './utils/addOnClickToCreationMenuItems';
@@ -22,10 +23,13 @@ import transformViewsItemsProps from './utils/transformViewsItemProps';
 
 const ACTIONS = {
 	createAsset: createAssetAction,
-	uploadMultipleFiles: multipleFilesUploadAction,
+	createFolder: createFolderAction,
 };
 
-export default function AllFDSPropsTransformer({
+const OBJECT_ENTRY_FOLDER_CLASSNAME =
+	'com.liferay.object.model.ObjectEntryFolder';
+
+export default function ContentFDSPropsTransformer({
 	additionalProps,
 	creationMenu,
 	itemsActions = [],
@@ -64,6 +68,11 @@ export default function AllFDSPropsTransformer({
 					type: 'internal',
 				} as IInternalRenderer,
 				{
+					component: SimpleActionLinkRenderer,
+					name: 'simpleActionLinkTableCellRenderer',
+					type: 'internal',
+				} as IInternalRenderer,
+				{
 					component: SpaceRenderer,
 					name: 'spaceTableCellRenderer',
 					type: 'internal',
@@ -75,25 +84,37 @@ export default function AllFDSPropsTransformer({
 				} as IInternalRenderer,
 			],
 		},
-		infoPanelComponent: () => AssetTypeInfoPanel({additionalProps}),
+		infoPanelComponent: (items: {items: ISearchAssetObjectEntry[]}) => (
+			<AssetTypeInfoPanel
+				additionalProps={additionalProps as any}
+				{...items}
+			/>
+		),
 		itemsActions: itemsActions.map((action) => {
-			if (action?.data?.id === 'download') {
+			if (action?.data?.id === 'actionLink') {
 				return {
 					...action,
 					isVisible: (item: any) =>
-						Boolean(item?.embedded?.file?.link?.href),
+						Boolean(
+							item?.entryClassName !==
+								OBJECT_ENTRY_FOLDER_CLASSNAME
+						),
 				};
 			}
 			else if (action?.data?.id === 'view-content') {
 				return {
 					...action,
-					isVisible: (item: any) => Boolean(!item?.embedded?.file),
+					isVisible: (item: any) =>
+						Boolean(
+							item?.entryClassName !==
+								OBJECT_ENTRY_FOLDER_CLASSNAME
+						),
 				};
 			}
 			else if (action?.data?.id === 'view-file') {
 				return {
 					...action,
-					isVisible: (item: any) => Boolean(item?.embedded?.file),
+					isVisible: () => false,
 				};
 			}
 
@@ -108,32 +129,7 @@ export default function AllFDSPropsTransformer({
 			event: Event;
 			itemData: any;
 		}) => {
-			if (action?.data?.id === 'show-details') {
-				Liferay.fire(EVENTS.ASSET_DATA, {items: [{...itemData}]});
-			}
-			else if (action?.data?.id === 'view-content') {
-				event?.preventDefault();
-
-				openModal({
-					size: 'full-screen',
-					title: itemData.embedded.title,
-					url: formatActionURL(itemData, action.href),
-				});
-			}
-			else if (action?.data?.id === 'view-file') {
-				openModal({
-					containerProps: {
-						className: '',
-					},
-					contentComponent: () =>
-						FilePreviewerModalContent({
-							file: itemData.embedded.file,
-							headerName: itemData.embedded.title,
-						}),
-					size: 'full-screen',
-				});
-			}
-			else if (action?.data?.id === 'share') {
+			if (action?.data?.id === 'share') {
 				const {autocompleteURL, collaboratorURLs} = additionalProps;
 
 				shareAction({
@@ -144,10 +140,15 @@ export default function AllFDSPropsTransformer({
 					title: itemData.embedded?.title,
 				});
 			}
-		},
+			else if (action?.data?.id === 'viewContent') {
+				event?.preventDefault();
 
-		onSelectedItemsChange: (selectedItems: any[]) => {
-			Liferay.fire(EVENTS.ASSET_DATA, {items: selectedItems});
+				openModal({
+					size: 'full-screen',
+					title: itemData.embedded.title,
+					url: formatActionURL(itemData, action.href),
+				});
+			}
 		},
 		views: transformViewsItemsProps(views),
 	};
