+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/props_transformer/FilesFDSPropsTransformer.tsx
@@ -5,12 +5,15 @@
 
 import {IInternalRenderer, IView} from '@liferay/frontend-data-set-web';
 import {openModal} from 'frontend-js-components-web';
+import React from 'react';
 
-import formatActionURL from '../../common/utils/formatActionURL';
+import {ISearchAssetObjectEntry} from '../../structure_builder/types/AssetType';
 import AssetTypeInfoPanel from '../info_panel/AssetTypeInfoPanelContent';
 import {EVENTS} from '../info_panel/util/constants';
+import FilePreviewerModalContent from '../modal/FilePreviewerModalContent';
 import createAssetAction from './actions/createAssetAction';
 import createFolderAction from './actions/createFolderAction';
+import multipleFilesUploadAction from './actions/multipleFilesUploadAction';
 import shareAction from './actions/shareAction';
 import AuthorRenderer from './cell_renderers/AuthorRenderer';
 import NameRenderer from './cell_renderers/NameRenderer';
@@ -23,12 +26,13 @@ import transformViewsItemsProps from './utils/transformViewsItemProps';
 const ACTIONS = {
 	createAsset: createAssetAction,
 	createFolder: createFolderAction,
+	uploadMultipleFiles: multipleFilesUploadAction,
 };
 
 const OBJECT_ENTRY_FOLDER_CLASSNAME =
 	'com.liferay.object.model.ObjectEntryFolder';
 
-export default function ContentFDSPropsTransformer({
+export default function FilesFDSPropsTransformer({
 	additionalProps,
 	creationMenu,
 	itemsActions = [],
@@ -83,9 +87,21 @@ export default function ContentFDSPropsTransformer({
 				} as IInternalRenderer,
 			],
 		},
-		infoPanelComponent: () => AssetTypeInfoPanel({additionalProps}),
+		infoPanelComponent: (items: {items: ISearchAssetObjectEntry[]}) => (
+			<AssetTypeInfoPanel
+				additionalProps={additionalProps as any}
+				{...items}
+			/>
+		),
 		itemsActions: itemsActions.map((action) => {
-			if (action?.data?.id === 'actionLink') {
+			if (action?.data?.id === 'download') {
+				return {
+					...action,
+					isVisible: (item: any) =>
+						Boolean(item?.embedded?.file?.link?.href),
+				};
+			}
+			else if (action?.data?.id === 'actionLink') {
 				return {
 					...action,
 					isVisible: (item: any) =>
@@ -98,6 +114,12 @@ export default function ContentFDSPropsTransformer({
 			else if (action?.data?.id === 'view-content') {
 				return {
 					...action,
+					isVisible: () => false,
+				};
+			}
+			else if (action?.data?.id === 'view-file') {
+				return {
+					...action,
 					isVisible: (item: any) =>
 						Boolean(
 							item?.entryClassName !==
@@ -105,25 +127,30 @@ export default function ContentFDSPropsTransformer({
 						),
 				};
 			}
-			else if (action?.data?.id === 'view-file') {
-				return {
-					...action,
-					isVisible: () => false,
-				};
-			}
 
 			return action;
 		}),
 		onActionDropdownItemClick: ({
 			action,
-			event,
 			itemData,
 		}: {
 			action: any;
-			event: Event;
 			itemData: any;
 		}) => {
-			if (action?.data?.id === 'share') {
+			if (action?.data?.id === 'view-file') {
+				openModal({
+					containerProps: {
+						className: '',
+					},
+					contentComponent: () =>
+						FilePreviewerModalContent({
+							file: itemData.embedded.file,
+							headerName: itemData.embedded.title,
+						}),
+					size: 'full-screen',
+				});
+			}
+			else if (action?.data?.id === 'share') {
 				const {autocompleteURL, collaboratorURLs} = additionalProps;
 
 				shareAction({
@@ -134,21 +161,6 @@ export default function ContentFDSPropsTransformer({
 					title: itemData.embedded?.title,
 				});
 			}
-			else if (action?.data?.id === 'show-details') {
-				Liferay.fire(EVENTS.ASSET_DATA, {items: [{...itemData}]});
-			}
-			else if (action?.data?.id === 'view-content') {
-				event?.preventDefault();
-
-				openModal({
-					size: 'full-screen',
-					title: itemData.embedded.title,
-					url: formatActionURL(itemData, action.href),
-				});
-			}
-		},
-		onSelectedItemsChange: (selectedItems: any[]) => {
-			Liferay.fire(EVENTS.ASSET_DATA, {items: selectedItems});
 		},
 		views: transformViewsItemsProps(views),
 	};
